+++ database/migrations/2016_06_01_000001_create_oauth_auth_codes_table.php
@@ -15,8 +15,8 @@ class CreateOauthAuthCodesTable extends Migration
     {
         Schema::create('oauth_auth_codes', function (Blueprint $table) {
             $table->string('id', 100)->primary();
-            $table->bigInteger('user_id');
-            $table->unsignedInteger('client_id');
+            $table->unsignedBigInteger('user_id')->index();
+            $table->unsignedBigInteger('client_id');
             $table->text('scopes')->nullable();
             $table->boolean('revoked');
             $table->dateTime('expires_at')->nullable();