+++ database/migrations/2016_06_01_000002_create_oauth_access_tokens_table.php
@@ -15,8 +15,8 @@ class CreateOauthAccessTokensTable extends Migration
     {
         Schema::create('oauth_access_tokens', function (Blueprint $table) {
             $table->string('id', 100)->primary();
-            $table->bigInteger('user_id')->index()->nullable();
-            $table->unsignedInteger('client_id');
+            $table->unsignedBigInteger('user_id')->nullable()->index();
+            $table->unsignedBigInteger('client_id');
             $table->string('name')->nullable();
             $table->text('scopes')->nullable();
             $table->boolean('revoked');