+++ database/migrations/2016_06_01_000003_create_oauth_refresh_tokens_table.php
@@ -15,7 +15,7 @@ class CreateOauthRefreshTokensTable extends Migration
     {
         Schema::create('oauth_refresh_tokens', function (Blueprint $table) {
             $table->string('id', 100)->primary();
-            $table->string('access_token_id', 100)->index();
+            $table->string('access_token_id', 100);
             $table->boolean('revoked');
             $table->dateTime('expires_at')->nullable();
         });