+++ src/Http/Controllers/ApproveAuthorizationController.php
@@ -36,6 +36,8 @@ class ApproveAuthorizationController
      */
     public function approve(Request $request)
     {
+        $this->assertValidAuthToken($request);
+
         $authRequest = $this->getAuthRequestFromSession($request);
 
         return $this->convertResponse(