+++ src/Http/Controllers/AuthorizationController.php
@@ -4,6 +4,7 @@ namespace Laravel\Passport\Http\Controllers;
 
 use Illuminate\Contracts\Routing\ResponseFactory;
 use Illuminate\Http\Request;
+use Illuminate\Support\Str;
 use Laminas\Diactoros\Response as Psr7Response;
 use Laravel\Passport\Bridge\User;
 use Laravel\Passport\ClientRepository;
@@ -73,6 +74,7 @@ class AuthorizationController
             return $this->approveRequest($authRequest, $user);
         }
 
+        $request->session()->put('authToken', $authToken = Str::random());
         $request->session()->put('authRequest', $authRequest);
 
         return $this->response->view('passport::authorize', [
@@ -80,6 +82,7 @@ class AuthorizationController
             'user' => $user,
             'scopes' => $scopes,
             'request' => $request,
+            'authToken' => $authToken,
         ]);
     }