+++ tests/AuthorizedAccessTokenControllerTest.php
@@ -5,6 +5,7 @@ namespace Laravel\Passport\Tests;
 use Illuminate\Http\Request;
 use Laravel\Passport\Client;
 use Laravel\Passport\Http\Controllers\AuthorizedAccessTokenController;
+use Laravel\Passport\RefreshTokenRepository;
 use Laravel\Passport\Token;
 use Laravel\Passport\TokenRepository;
 use Mockery as m;
@@ -19,6 +20,11 @@ class AuthorizedAccessTokenControllerTest extends TestCase
     protected $tokenRepository;
 
     /**
+     * @var \Mockery\Mock|\Laravel\Passport\RefreshTokenRepository
+     */
+    protected $refreshTokenRepository;
+
+    /**
      * @var AuthorizedAccessTokenController
      */
     protected $controller;
@@ -26,7 +32,8 @@ class AuthorizedAccessTokenControllerTest extends TestCase
     protected function setUp(): void
     {
         $this->tokenRepository = m::mock(TokenRepository::class);
-        $this->controller = new AuthorizedAccessTokenController($this->tokenRepository);
+        $this->refreshTokenRepository = m::mock(RefreshTokenRepository::class);
+        $this->controller = new AuthorizedAccessTokenController($this->tokenRepository, $this->refreshTokenRepository);
     }
 
     protected function tearDown(): void
@@ -78,6 +85,7 @@ class AuthorizedAccessTokenControllerTest extends TestCase
         $token1->shouldReceive('revoke')->once();
 
         $this->tokenRepository->shouldReceive('findForUser')->andReturn($token1);
+        $this->refreshTokenRepository->shouldReceive('revokeRefreshTokensByAccessTokenId')->once();
 
         $request->setUserResolver(function () {
             $user = m::mock();