+++ CHANGELOG.md
@@ -1,6 +1,30 @@
 # Release Notes
 
-## [Unreleased](https://github.com/laravel/passport/compare/v8.2.0...master)
+## [Unreleased](https://github.com/laravel/passport/compare/v8.4.0...master)
+
+
+## [v8.4.0 (2020-02-12)](https://github.com/laravel/passport/compare/v8.3.1...v8.4.0)
+
+### Changed
+- Implement auth token for access requests ([#1188](https://github.com/laravel/passport/pull/1188))
+
+### Fixed
+- Revoke refresh tokens when auth tokens get revoked ([#1186](https://github.com/laravel/passport/pull/1186))
+
+
+## [v8.3.1 (2020-01-29)](https://github.com/laravel/passport/compare/v8.3.0...v8.3.1)
+
+### Fixed
+- Remove foreign keys ([20e9b66](https://github.com/laravel/passport/commit/20e9b66fcd003ba41301fc5de23b9892e307051a))
+
+
+## [v8.3.0 (2020-01-28)](https://github.com/laravel/passport/compare/v8.2.0...v8.3.0)
+
+### Added
+- Add a Passport Client factory to Passport publishing ([#1171](https://github.com/laravel/passport/pull/1171))
+
+### Changed
+- Use bigIncrements and indexes on relationships ([#1169](https://github.com/laravel/passport/pull/1169), [140a693](https://github.com/laravel/passport/commit/140a693a079f5611b3342360cde00b10e94162c1))
 
 
 ## [v8.2.0 (2020-01-07)](https://github.com/laravel/passport/compare/v8.1.0...v8.2.0)